Finished converting process_backslash_commands() to new system
